Cleaned a bit for pep8 and updated to latest pyomeca
