+++ .vscode/settings.json
@@ -2,12 +2,10 @@
     "python.testing.pytestArgs": ["tests"],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true,
-    "python.analysis.extraPaths": ["./external/pyScienceMode"],
-    "python.envFile": "${workspaceFolder}/.vscode/.env",
     "[python]": {
         "editor.defaultFormatter": "ms-python.black-formatter",
         "editor.formatOnSave": true,
-        "editor.rulers": [120]
+        "editor.rulers": [120],
     },
-    "cmake.configureOnOpen": false
+    "cmake.configureOnOpen": false,
 }
