+++ tests/test_Gui.py
@@ -1,15 +1,15 @@
-import pathlib
+import os
 
 import biorbd
 from bioviz import Viz
 
 
-def get_test_folder():
-    return pathlib.Path(__file__).parent.resolve()
+def get_base_folder():
+    return f"{os.getcwd()}/../"
 
 
 def test_model_load():
-    model_path = f"{get_test_folder()}/../examples/pyomecaman.bioMod"
+    model_path = f"{get_base_folder()}/examples/pyomecaman.bioMod"
 
     # From path
     b1 = Viz(model_path=model_path)
