--- testsTravis/test_noGui.py
@@ -1,5 +1,0 @@
-from bioviz import Viz
-
-
-def test_model_load():
-    pass
